+++ src/utils/searchElement.js
@@ -1,34 +1,36 @@
 import { getEdgeUniqueId, getElementProperties } from '../constants/functions'
 import { LOW_LEVEL_PROPERTIES } from '../constants/graph'
+import store from '../store'
 
 /**
  * Search free-text in elements' properties
  * @param  {Object}   params
- * @param  {Object}   params.classesFromApi          Nodes from initial OwlClasses
- * @param  {Array}    params.edgesIdsToDisplay       Array of edges IDs to display
- * @param  {Array}    params.nodesIdsToDisplay       Array of nodes IDs to display
- * @param  {Object}   params.objectPropertiesFromApi Edges from initial OwlObjectProperties
  * @param  {String}   params.search                  Search string
  * @param  {Function} params.setStoreState           setStoreState action
  * @return { undefined }
  */
 const searchElement = ({
-  classesFromApi,
-  edgesIdsToDisplay,
-  nodesIdsToDisplay,
-  objectPropertiesFromApi,
   search,
   setStoreState
 }) => {
+  const {
+    classesFromApi,
+    edgesIdsToDisplay,
+    objectPropertiesFromApi,
+    availableNodesNormalised
+  } = store.getState()
+
   const elementsToDisplay = {}
 
   if (search === '') {
     return setStoreState('freeTextSelection', JSON.parse(JSON.stringify(elementsToDisplay)))
   }
 
-  if (nodesIdsToDisplay.length > 0) {
-    for (let index = 0; index < nodesIdsToDisplay.length; index++) {
-      const nodeId = nodesIdsToDisplay[index]
+  const availableNodesNormalisedNodeIds = Object.keys(availableNodesNormalised)
+
+  if (availableNodesNormalisedNodeIds.length > 0) {
+    for (let index = 0; index < availableNodesNormalisedNodeIds.length; index++) {
+      const nodeId = availableNodesNormalisedNodeIds[index]
       const nodeElement = classesFromApi[nodeId]
 
       // check content in low level properties